requirements.txt
APScheduler==3.9.1
Flask_APScheduler==1.12.3
PyYAML==6.0

config~=0.5.1
Flask~=2.1.2
requests~=2.27.1
crypto~=1.4.1
utils~=1.0.1